+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Add
+- Add `IMGPROXY_OPEN_TELEMETRY_GRPC_INSECURE` config.
 
 ## [3.10.0] - 2022-11-04
 ### Add